wip: move button to sidebar
