+++ packages/frontend/src/components/Explorer/ExplorerHeader/index.tsx
@@ -5,7 +5,6 @@ import { useExplorerContext } from '../../../providers/ExplorerProvider';
 import { Can } from '../../common/Authorization';
 import ShareShortLinkButton from '../../common/ShareShortLinkButton';
 import ExploreFromHereButton from '../../ExploreFromHereButton';
-import { RefreshButton } from '../../RefreshButton';
 import RefreshDbtButton from '../../RefreshDbtButton';
 import SaveChartButton from '../SaveChartButton';
 
@@ -44,7 +43,6 @@ const ExplorerHeader: FC = memo(() => {
                     <RefreshDbtButton />
                 </Box>
                 <Group spacing="xs">
-                    <RefreshButton />
                     {!savedChart && (
                         <Can I="manage" a="SavedChart">
                             <SaveChartButton isExplorer />
